+++ OpenSpeed/Game.MW05/Types/UTL/_ListSet.h
@@ -22,7 +22,7 @@
 #include <OpenSpeed/Game.MW05/Types/UTL/List.h>
 
 namespace OpenSpeed::MW05::UTL {
-  template <typename T, std::uint32_t nT, typename E, std::uint32_t nE>
+  template <typename T, std::size_t nT, typename E, std::size_t nE>
   struct _ListSet {
     List<T, nT> _buckets[nE];
